completed models. added mapper.

+++ ImgurNet/Models/Account.cs
@@ -2,11 +2,11 @@
 {
     public class Account
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Url { get; set; }
         public string Bio { get; set; }
         public float Reputation { get; set; }
-        public int Created { get; set; }
+        public long Created { get; set; }
         public bool ProExpiration { get; set; }
     }
 }
